All numbers are now doubles in backend (to allow for decimal places in dkp), truncate decay to two decimal places, nicer loading of raider detail page

+++ Insomnia.Core/Controllers/RaiderController.cs
@@ -31,7 +31,7 @@ namespace Insomnia.Core.Controllers
         [Authorize]
         [HttpPost]
         [Route("multiple/decay")]
-        public async Task<Raider[]> DecayRaiders([FromBody] decimal percentage)
+        public async Task<Raider[]> DecayRaiders([FromBody] double percentage)
         {
             return await _raiderService.DecayRaiders(percentage);
         }
