+++ Insomnia.Core/Services/IRaiderService.cs
@@ -7,7 +7,7 @@ namespace Insomnia.Core.Services
     {
         Task<Raider> InsertRaider(RaiderEntity raider);
         Task<Raider[]> GetRaiders();
-        Task<Raider[]> DecayRaiders(decimal percentage);
+        Task<Raider[]> DecayRaiders(double percentage);
         Task<Raider[]> AddDkpToRaiders(AddDkpToRaiders raidersAndDkp);
         Task<Raider> GetRaider(string characterClass, string name);
     }
